Tk newgame smartcity addclientsocket

+++ components/socket/server/src/Tepuy/SocketSessions.class.php
@@ -11,13 +11,16 @@ use Tepuy\AppCodeException;
 class SocketSessions {
     private static $_sessions = array();
     private static $_resources = array();
+    private static $_settings = array();
+    private static $_settingsbyuid = array();
 
-    public static function addConnection($conn, $sess) {
+    public static function addConnection($conn, $sess, $iscron = false) {
 
         if (!isset(self::$_sessions[$sess->uid])) {
             self::$_sessions[$sess->uid] = new \stdClass();
             self::$_sessions[$sess->uid]->clients = new \SplObjectStorage;
             self::$_sessions[$sess->uid]->skeys = array();
+            self::$_sessions[$sess->uid]->crons = array();
         }
 
         self::$_resources[$conn->resourceId] = $sess->uid;
@@ -25,12 +28,27 @@ class SocketSessions {
         // Store the new connection to send messages to later
         self::$_sessions[$sess->uid]->clients->attach($conn);
         self::$_sessions[$sess->uid]->skeys[$conn->resourceId] = $sess;
+
+        if ($iscron) {
+            self::$_sessions[$sess->uid]->crons[$conn->resourceId] = true;
+        }
     }
 
     public static function rmConnection($conn) {
         $session = self::getByResourceId($conn->resourceId);
-        $session->clients->detach($conn);
-        unset(self::$_resources[$conn->resourceId]);
+        if ($session) {
+            $session->clients->detach($conn);
+        }
+
+        if (isset(self::$_resources[$conn->resourceId])) {
+            unset(self::$_resources[$conn->resourceId]);
+        }
+
+        if (isset($session->crons[$conn->resourceId])) {
+            unset($session->crons[$conn->resourceId]);
+        }
+
+
     }
 
     public static function isActiveSessKey($id) {
@@ -40,6 +58,11 @@ class SocketSessions {
         }
 
         return isset($session->skeys[$id]);
+    }
+
+    public static function isCron($id) {
+        $session = self::getByResourceId($id);
+        return isset($session->crons[$id]);
     }
 
     public static function countClients($id) {
@@ -60,6 +83,54 @@ class SocketSessions {
     public static function getClientsById($id) {
         $session = self::getByResourceId($id);
         return !isset($session) ? null : $session->clients;
+    }
+
+    public static function getUIDById($id) {
+        if (!isset(self::$_resources[$id])) {
+            return null;
+        }
+
+        return self::$_resources[$id];
+    }
+
+    public static function setSettings($settings) {
+
+        self::$_settings = $settings;
+
+        self::$_settingsbyuid = array();
+        foreach($settings as $setting) {
+            $settingsdata = json_decode($setting->param1);
+            self::$_settingsbyuid[$setting->uid] = $settingsdata;
+        }
+    }
+
+    public static function getGameActions($id) {
+
+        $actions = self::getGameProperty($id, 'actions');
+
+        return !$actions ? array() : $actions;
+    }
+
+    public static function getGameKey($id) {
+        return self::getGameProperty($id, 'game');
+    }
+
+    private static function getGameProperty($id, $prop) {
+
+        if (!isset(self::$_resources[$id])) {
+            return null;
+        }
+
+        $resourceid = self::$_resources[$id];
+
+        if (isset(self::$_settingsbyuid[$resourceid])) {
+            $params = self::$_settingsbyuid[$resourceid];
+            if (property_exists($params, $prop)) {
+                return $params->$prop;
+            }
+        }
+
+        return null;
     }
 
     private static function getByResourceId($id) {
